Добавлена проверка на удаление сущностей, принадлежащих транзакциям

+++ src/main/java/ru/fess38/finance/controller/AccountController.java
@@ -8,6 +8,7 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.RestController;
 import ru.fess38.finance.dao.AccountDao;
+import ru.fess38.finance.dao.TransactionDao;
 import ru.fess38.finance.model.Account;
 import ru.fess38.finance.model.ModifiableAccount;
 
@@ -17,6 +18,8 @@ import java.util.List;
 public class AccountController {
   @Autowired
   private AccountDao accountDao;
+  @Autowired
+  private TransactionDao transactionDao;
 
   @RequestMapping(value = "/account/get", method = RequestMethod.GET)
   public @ResponseBody List<Account> get() {
@@ -45,6 +48,8 @@ public class AccountController {
 
   @RequestMapping(value = "/account/delete", method = RequestMethod.POST)
   public void delete(@RequestBody Account account) {
-    accountDao.delete(account);
+    if (transactionDao.countByAccount(account) == 0) {
+      accountDao.delete(account);
+    }
   }
 }
