+++ src/main/java/ru/fess38/finance/controller/RubricController.java
@@ -8,6 +8,7 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.RestController;
 import ru.fess38.finance.dao.RubricDao;
+import ru.fess38.finance.dao.TransactionDao;
 import ru.fess38.finance.model.ModifiableRubric;
 import ru.fess38.finance.model.Rubric;
 
@@ -17,6 +18,8 @@ import java.util.List;
 public class RubricController {
   @Autowired
   private RubricDao rubricDao;
+  @Autowired
+  private TransactionDao transactionDao;
 
   @RequestMapping(value = "/rubric/get", method = RequestMethod.GET)
   public @ResponseBody List<Rubric> get() {
@@ -40,6 +43,8 @@ public class RubricController {
 
   @RequestMapping(value = "/rubric/delete", method = RequestMethod.POST)
   public void delete(@RequestBody Rubric rubric) {
-    rubricDao.delete(rubric);
+    if (transactionDao.countByRubric(rubric) == 0) {
+      rubricDao.delete(rubric);
+    }
   }
 }
