+++ src/main/java/ru/fess38/finance/controller/TagController.java
@@ -8,6 +8,7 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.RestController;
 import ru.fess38.finance.dao.TagDao;
+import ru.fess38.finance.dao.TransactionDao;
 import ru.fess38.finance.model.ModifiableTag;
 import ru.fess38.finance.model.Tag;
 
@@ -17,6 +18,8 @@ import java.util.List;
 public class TagController {
   @Autowired
   private TagDao tagDao;
+  @Autowired
+  private TransactionDao transactionDao;
 
   @RequestMapping(value = "/tag/get", method = RequestMethod.GET)
   public @ResponseBody List<Tag> get() {
@@ -35,6 +38,8 @@ public class TagController {
 
   @RequestMapping(value = "/tag/delete", method = RequestMethod.POST)
   public void delete(@RequestBody Tag tag) {
-    tagDao.delete(tag);
+    if (transactionDao.countByTag(tag) == 0) {
+      tagDao.delete(tag);
+    }
   }
 }
