+++ src/main/java/ru/fess38/finance/controller/UserController.java
@@ -7,6 +7,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.RestController;
+import ru.fess38.finance.dao.TransactionDao;
 import ru.fess38.finance.dao.UserDao;
 import ru.fess38.finance.model.ModifiableUser;
 import ru.fess38.finance.model.User;
@@ -17,6 +18,8 @@ import java.util.List;
 public class UserController {
   @Autowired
   private UserDao userDao;
+  @Autowired
+  private TransactionDao transactionDao;
 
   @RequestMapping(value = "/user/get", method = RequestMethod.GET)
   public @ResponseBody List<User> get() {
@@ -35,6 +38,8 @@ public class UserController {
 
   @RequestMapping(value = "/user/delete", method = RequestMethod.POST)
   public void delete(@RequestBody User user) {
-    userDao.delete(user);
+    if (transactionDao.countByUser(user) == 0) {
+      userDao.delete(user);
+    }
   }
 }
